+++ servers/lib/servicesCenter.js
@@ -15,8 +15,9 @@ const {resolve} = require("path");
 const {throwNoThisFunction, throwNoThisServer, throwOperationFailed} = require(resolve(__dirname, "..", "..", "errors"));
 const services = resolve(resolve(__dirname, "..", "..", "services"));
 
-const [workflowService, userService, goodsService, tagsService] = [
-  require(resolve(services, "workflowService")), require(resolve(services, "userService")), require(resolve(services, "goodsService")), require(resolve(services, "tagsService"))];
+const [workflowService, userService, goodsService, tagsService, {systemInfo, hardDiskInfo, memoryInfo}] = [
+  require(resolve(services, "workflowService")), require(resolve(services, "userService")), require(resolve(services, "goodsService")),
+  require(resolve(services, "tagsService")), require(resolve(services, "fileSystem"))];
 
 const REFISTRY = Symbol("REFISTRY");
 module[REFISTRY] = new Map();
@@ -25,6 +26,9 @@ module[REFISTRY].set("workflowService", workflowService);
 module[REFISTRY].set("userService", userService);
 module[REFISTRY].set("goodsService", goodsService);
 module[REFISTRY].set("tagsService", tagsService);
+module[REFISTRY].set("systemService", {
+  systemInfo, hardDiskInfo, memoryInfo
+});
 
 /**
  * 获取执行方法
@@ -38,7 +42,7 @@ const getService = (serviceName, funcName, user) => {
   if (!user) {
     throwOperationFailed();
   }
-  
+
   const service = module[REFISTRY].get(serviceName);
   if (!service) {
     throwNoThisServer();